requirements.txt
python-dotenv
pyrogram
TgCrypto